feat: 주요기록 캐러셀을 카드로 대체(#77)
